refactor(backup): co-locate restic message parsers in new restic module
Moves ResticMessage / ResticStatus / ResticSummary and
parse_restic_message from src/output.rs to a new
src/services/backup/restic.rs. Restic types are an implementation
detail of BackupSession (push/prune); they should not leak into
output.rs which is a CLI display utility.
