+++ src/services/backup.rs
@@ -1,3 +1,4 @@
 pub mod executor;
 pub mod recipe;
+pub mod restic;
 pub mod ssh;
